Add a probe that reads the season, and apply rustfmt to the weather probes
The season probe reports the degrees the sun gives at latitudes from the
equator to the pole, at midsummer and at midwinter, and counts the rows
that see a swing above a stated amount.

  cargo run -p cachette-core --release --example weather_season_probe \
    -- 256 0 60

  row  latitude   midsummer  midwinter  swing  mean
    0      -128         -80         38    118   -21
   32       -96         -69         68    137     0
  128         0          38         38      0    38
  224        96          68        -69    137     0
  255       127          39        -80    119   -20

  rows with a swing of 60 or more: 201 of 256

Neither amplitude should change, and the measurements say why.

The tilt carries the sun half way to a pole. Raising it to two thirds
warms the annual mean of a pole from -21 to -8 and cools the equator from
38 to 21, so the gradient from pole to equator falls from 58 degrees to
29. That halves the cold poles, which is the feature this design earned
from one term. It buys almost nothing: the rows that see a swing of 60 or
more move from 201 to 199. Four fifths of the map already sees a real
season, so the tilt is not the limit. The boundary sits below two thirds
and the current half is about right.

The seasonal swing cannot rise without taking from something. A build-time
assertion holds the base equal to the season swing plus the cloud swing,
and a second holds the four terms to span the heat scale exactly. The
other three terms all now do work a watcher sees: the ground term carries
the depth grading that put air over land, and the cloud term couples the
water back into the temperature. Widening the scale changes no ratio, so
it changes no picture.

The field already swings hard. Over one season period the mean temperature
of the pole row runs 44, 87, 163, 87 on a scale of 256, a swing of 119:

  cargo run -p cachette-core --release --example weather_diagnosis_probe \
    -- 128 47 2048 1

The season is not what was undramatic. The clouds were.

+++ crates/cachette-core/examples/weather_rotation_probe.rs
@@ -136,7 +136,7 @@ fn distance(from: (i64, i64), to: (i64, i64)) -> i64 {
 fn main() {
     let extent = argument(1, 256) as u32;
     let seed = argument(2, 0x2f);
-    let ticks = argument(3, 400) as u64;
+    let ticks = argument(3, 400);
     let bits = argument(4, 0) as u32;
     let scale = WeatherScale::from_bits(bits).expect("the scale describes a lattice");
     let mut world = World::with_weather_scale(
@@ -162,9 +162,7 @@ fn main() {
 
     let ceiling = circulation_ceiling();
     println!("extent {extent} seed {seed:#x} ticks {ticks} scale bits {bits}");
-    println!(
-        "circulation ceiling {ceiling}, storm mark {mark} drops, storm floor {floor} cells"
-    );
+    println!("circulation ceiling {ceiling}, storm mark {mark} drops, storm floor {floor} cells");
 
     let mut live: Vec<Tracked> = Vec::new();
     let mut done: Vec<Tracked> = Vec::new();
